changed route of new home page
