+++ src/cms/preview-templates/IndexPagePreview.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { IndexPageTemplate } from '../../templates/index-page'
+import { HomePageTemplate } from '../../templates/index-page';
 
 const IndexPagePreview = ({ entry, getAsset }) => {
   const data = entry.getIn(['data']).toJS();
   if (data) {
     return (
-      <IndexPageTemplate
+      <HomePageTemplate
         data={data}
       />
     )
