+++ src/pages/homeLegacy/index.md
@@ -1,0 +1,88 @@
+---
+templateKey: home-legacy
+title: Samagra | Transforming Governance
+description: Samagra | Transforming Governance
+stickyBanner:
+  stickyBannerActive: false
+  stickyBannerButtonColor: orange
+  stickyBannerButtonLink: https://www.samagragovernance.in/amritseries
+  stickyBannerButtonTextColor: white
+  stickyBannerButtonTitle: here
+  stickyBannerColor: white
+  stickyBannerTitle: Check out our newly launched Amrit Series
+  stickyBannerTitleColor: black
+media:
+  - mediaItem: "1 Stop, 556 Schemes: Haryana Govt’s Unique System Can Revolutionise
+      How India Works"
+baseBanner:
+  titleLines:
+    - text: Transforming
+    - text: Governance
+subBanners:
+  - projectName: NIPUN Bharat cell
+    color: '#BA9807'
+    slides:
+      - image: /img/NIPUN-Bharat-slide-1.png
+        logo: /img/nipun-logo.png
+        title: '1,10,000 Primary Schools | 60,00,000 Students'
+    titleLines:
+      - text: Transforming
+      - text: Education
+  - projectName: Antyodaya Saral
+    color: '#76323F'
+    slides:
+      - image: /img/antyodaya-Saral-slide-1.png
+        logo: /img/Digital-Haryana-Logo.png
+        title: '550 Services | 2.7 Cr Citizens'
+    titleLines:
+      - text: Transforming
+      - text: Service Delivery
+  - projectName: KONNECT
+    projectName2: LEAP
+    color: '#428F37'
+    slides:
+      - image: /img/KONNECT-LEAP-slide-1.jpeg
+        logo: /img/KO-LEAP-logo.png
+        title: '100+ Ecosystem Partners | 75,00,000 Farmers'
+    titleLines:
+      - text: Transforming
+      - text: Agriculture
+  - projectName: Anamaya
+    color: '#027F50'
+    slides:
+      - image: /img/STRIDE-slide-1.jpg
+        logo: /img/Niramaya-logo.png
+        title: '650 Institutes | 50,000 Students'
+    titleLines:
+      - text: Transforming
+      - text: Health
+secondSection:
+  ourApproach:
+    description:
+      - subTitle: Our Approach
+        text: We combine a top-down management consulting approach for problem
+          structuring with a bottom-up understanding of the governance ecosystem
+          for designing solutions, while leveraging data and technology to
+          enable implementation of our solutions, with the objective of making
+          governments accountable and deliver with minimum delays and maximum
+          efficiency.
+    image: /img/revised-approach_high-res.png
+    title: Our Approach
+  ourModel:
+    image: /img/home-our-vision.jpg
+    title: Our Model
+    description:
+      - subTitle: Who we are 
+        text: We are a mission-driven governance consulting firm. We firmly believe in governance being the primary lever of change to create large scale impact in the country. 
+      - subTitle: What we do
+        text: We work with the senior political and bureaucratic leadership of states to solve governance problems at scale. We co-work with the government to diagnose the problem, design a transformation roadmap and implement the same. Our solutions are rooted in the realities of governance in India. We leverage tech & data to enable systemic transformations. 
+
+  title: We are working to improve the quality of life of citizens through better
+    governance
+
+thirdSection:
+  img1: /img/saral-header-bg.jpg
+  img2: /img/Agriculture.jpg
+  img3: /img/STRIDE-slide-1.jpg
+  img4: /img/Education.jpg
+---
