+++ src/pages/index.md
@@ -18,71 +18,112 @@ baseBanner:
   titleLines:
     - text: Transforming
     - text: Governance
-subBanners:
-  - projectName: NIPUN Bharat cell
-    color: '#BA9807'
-    slides:
-      - image: /img/NIPUN-Bharat-slide-1.png
-        logo: /img/nipun-logo.png
-        title: '1,10,000 Primary Schools | 60,00,000 Students'
-    titleLines:
-      - text: Transforming
-      - text: Education
-  - projectName: Antyodaya Saral
-    color: '#76323F'
-    slides:
-      - image: /img/antyodaya-Saral-slide-1.png
-        logo: /img/Digital-Haryana-Logo.png
-        title: '550 Services | 2.7 Cr Citizens'
-    titleLines:
-      - text: Transforming
-      - text: Service Delivery
-  - projectName: KONNECT
-    projectName2: LEAP
-    color: '#428F37'
-    slides:
-      - image: /img/KONNECT-LEAP-slide-1.jpeg
-        logo: /img/KO-LEAP-logo.png
-        title: '100+ Ecosystem Partners | 75,00,000 Farmers'
+baseBannerVideo: /img/home_updated_video.mp4
+ourMission:
+  title: Our Mission
+  description: We are working to improve the quality of life of citizens through
+    better governance
+ourModel:
+  title: Our Model
+  description:
+    - subTitle: Who we are
+      text: We are a mission-driven governance consulting firm. We firmly believe in
+        governance being the primary lever of change to create large scale
+        impact in the country.
+    - subTitle: What we do
+      text: We work with the senior political and bureaucratic leadership of states to
+        solve governance problems at scale. We co-work with the government to
+        diagnose the problem, design a transformation roadmap and implement the
+        same. Our solutions are rooted in the realities of governance in India.
+        We leverage tech & data to enable systemic transformations.
+  ourModelVideoLink: https://www.youtube.com/embed/Y-iHVVfSHRg
+ourApproach:
+  title: Our Approach
+  description: We combine a top-down management consulting approach for problem
+    structuring with a bottom-up understanding of the governance ecosystem for
+    designing solutions, while leveraging data and technology to enable
+    implementation of our solutions, with the objective of making governments
+    accountable and deliver with minimum delays and maximum efficiency.
+  approachVideo: /img/approach_view.mp4
+ourWorkDescription: We work with state governments across domains
+ourWork:
+  - projectName:
+      - project: KONNECT
+      - project: LEAP
     titleLines:
       - text: Transforming
       - text: Agriculture
-  - projectName: Anamaya
-    color: '#027F50'
-    slides:
-      - image: /img/STRIDE-slide-1.jpg
-        logo: /img/Niramaya-logo.png
-        title: '650 Institutes | 50,000 Students'
+    description1: Our work in the agriculture domain spans 7 years which has
+      impacted the lives of 3 Cr+ farmers across the states of Odisha
+      and Uttar Pradesh
+    image: /img/agriculture.png
+    workLogo:
+      - logo: /img/konnect-logo.png
+      - logo: /img/leap-logo-copy.png
+    backgroundMap: /img/odisha-map.png
+    description2: Increasing farmer income via phased cohort-based enhancement
+      approach for 75 lakh farmers in Odisha
+  - projectName:
+      - project: NIPUN Bharat cell
+    titleLines:
+      - text: Transforming
+      - text: Education
+    description1: Our work in the education domain spans 7 years which has impacted
+      the lives of 1 Cr+ students across the states of Haryana, Himachal Pradesh
+      and Uttar Pradesh
+    image: /img/education.png
+    workLogo:
+      - logo: /img/nipun-logo-copy.png
+    backgroundMap: /img/up-map.png
+    description2: Enabling 60 lakh students in grades 1-3 to attain foundational
+      literacy & numeracy in UP
+  - projectName:
+      - project: Anamaya
     titleLines:
       - text: Transforming
       - text: Health
-secondSection:
-  ourApproach:
-    description:
-      - subTitle: Our Approach
-        text: We combine a top-down management consulting approach for problem
-          structuring with a bottom-up understanding of the governance ecosystem
-          for designing solutions, while leveraging data and technology to
-          enable implementation of our solutions, with the objective of making
-          governments accountable and deliver with minimum delays and maximum
-          efficiency.
-    image: /img/revised-approach_high-res.png
-    title: Our Approach
-  ourModel:
-    image: /img/home-our-vision.jpg
-    title: Our Model
-    description:
-      - subTitle: Who we are 
-        text: We are a mission-driven governance consulting firm. We firmly believe in governance being the primary lever of change to create large scale impact in the country. 
-      - subTitle: What we do
-        text: We work with the senior political and bureaucratic leadership of states to solve governance problems at scale. We co-work with the government to diagnose the problem, design a transformation roadmap and implement the same. Our solutions are rooted in the realities of governance in India. We leverage tech & data to enable systemic transformations. 
-
-  title: We are working to improve the quality of life of citizens through better
-    governance
-
-thirdSection:
-  img1: /img/saral-header-bg.jpg
-  img2: /img/Agriculture.jpg
-  img3: /img/STRIDE-slide-1.jpg
-  img4: /img/Education.jpg
+    description1: Our work in the health domain spans 2 years which has impacted the
+      lives of 1 Cr+ citizens across the states of Odisha and Uttar Pradesh
+    image: /img/health.png
+    workLogo:
+      - logo: /img/stride-logo.png
+    backgroundMap: /img/odisha-map.png
+    description2: Establishing systemic enablers for improvement in health &
+      nutrition outcomes of 1 Cr+ tribals in Odisha
+  - projectName:
+      - project: Antyodaya Saral
+    titleLines:
+      - text: Transforming
+      - text: Service Delivery
+    description1: Our work in the service delivery domain spans 7 years which has
+      impacted the lives of 2.7 Cr+ citizens in the state of Haryana
+    image: /img/service-delivery.png
+    workLogo:
+      - logo: /img/saral-logo-copy.png
+    backgroundMap: /img/haryana-map.png
+    description2: Transforming scheme and service delivery for 2.7 Cr citizens in Haryana
+ourJourney:
+  - subHeading: 15+ Programs
+    description: Across geographies and domains
+    image: /img/programs.png
+  - subHeading: 7 Domains
+    description: Our impactful work spans across domains like health, education,
+      agriculture, skilling, employment and public service delivery among others
+    image: /img/domains.png
+  - subHeading: 7 Governments
+    description: Empowered by a track record of partnership with governments
+      (including state & central)
+    image: /img/government.png
+  - subHeading: 125+ Team
+    description: Dynamic team of young professionals, each driven by a passion for
+      innovation and excellence
+    image: /img/team.png
+  - subHeading: 25 Implementation Partners
+    description: We collaborate with technical experts to execute quality
+      implementation at scale
+    image: /img/implementation_partner.png
+  - subHeading: 50 mn Lives Impacted
+    description: In collaboration with governments and partners, we have transformed
+      the lives of farmers, students, teachers and many more
+    image: /img/live_impacted.png
 ---
